Move pipelines to dedicated package, refactor MultiDistantMeasure
