+++ eradiate/_util.py
@@ -204,3 +204,9 @@ def deduplicate(value: t.Sequence, preserve_order: bool = True) -> t.List:
 
     else:
         return list(set(value))
+
+
+def camel_to_snake(name):
+    # from https://stackoverflow.com/questions/1175208/elegant-python-function-to-convert-camelcase-to-snake-case
+    name = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", name)
+    return re.sub("([a-z0-9])([A-Z])", r"\1_\2", name).lower()
