+++ eradiate/pipelines/_compute.py
@@ -1,0 +1,12 @@
+import attr
+
+from ._core import PipelineStep
+from ..attrs import parse_docs
+
+
+@parse_docs
+@attr.s
+class ComputeReflectance(PipelineStep):
+    """
+    Compute reflectance data.
+    """
